Added station ID, logo ID and image token + authentication for POST, PUT and DELETE methods

+++ app/api/api_v1/endpoints/countries.py
@@ -29,12 +29,18 @@ def get_countries() -> Any:
 # -----------------------------------------------------------
 @router.post("/create/", response_model=schemas.Country)
 def create_country(
-        country_in: schemas.Country
+        country_in: schemas.Country,
+        current_user: models.User = Depends(deps.get_current_active_user)
 ) -> Any:
-    db = next(deps.get_db())
-    new_country = crud.country.create_country(db,
-                                              country=country_in)
-    return new_country
+    if not crud.user.is_superuser(current_user):
+        raise HTTPException(
+            status_code=400,
+            detail="The user doesn't have enough privileges"
+        )
+    else:
+        db = next(deps.get_db())
+        country = crud.country.create_country(db, country_in)
+        return country
 
 
 # UPDATE
@@ -42,25 +48,39 @@ def create_country(
 @router.put("/update/", response_model=schemas.Country)
 def update_country(
         country_code: str,
-        country_in: schemas.Country
+        country_in: schemas.Country,
+        current_user: models.User = Depends(deps.get_current_active_user)
 ):
-    db = next(deps.get_db())
-    country = crud.country.get_country(db, country_code)
-    if not country:
-        raise HTTPException(status_code=404, detail="Country not found")
-    country = crud.country.update_country(db, country_code, country_in)
-    return country
+    if not crud.user.is_superuser(current_user):
+        raise HTTPException(
+            status_code=400,
+            detail="The user doesn't have enough privileges"
+        )
+    else:
+        db = next(deps.get_db())
+        country = crud.country.get_country(db, country_code)
+        if not country:
+            raise HTTPException(status_code=404, detail="Country not found")
+        country = crud.country.update_country(db, country_code, country_in)
+        return country
 
 
 # DELETE
 # -----------------------------------------------------------
 @router.delete("/delete/", response_model=schemas.Country)
 def delete_country(
-        country_code: str
+        country_code: str,
+        current_user: models.User = Depends(deps.get_current_active_user)
 ):
-    db = next(deps.get_db())
-    country = crud.country.get_country(db, country_code)
-    if not country:
-        raise HTTPException(status_code=404, detail="Country not found")
-    country = crud.country.delete_country(db, country_code)
-    return country
+    if not crud.user.is_superuser(current_user):
+        raise HTTPException(
+            status_code=400,
+            detail="The user doesn't have enough privileges"
+        )
+    else:
+        db = next(deps.get_db())
+        country = crud.country.get_country(db, country_code)
+        if not country:
+            raise HTTPException(status_code=404, detail="Country not found")
+        country = crud.country.delete_country(db, country_code)
+        return country
